<refactor> 패키징 및 클래스명 수정
