+++ src/main/kotlin/model/generator/NumberGenerator.kt
@@ -1,4 +1,4 @@
-package data.generator
+package model.generator
 
 interface NumberGenerator {
     fun generate(): Int
