+++ src/test/kotlin/domain/RefereeTest.kt
@@ -1,21 +1,23 @@
 package domain
 
-import data.Car
+import model.Car
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 
-class CarRacingGameTest {
+class RefereeTest {
+
+    private val referee: Referee = Referee()
 
     @Test
     fun `우승자 정하기`() {
         val cars = listOf(Car(NAME1, WINNER_POSITION), Car(NAME2, LOSER_POSITION))
-        assertThat(CarRacingGame().decideWinner(cars)).isEqualTo(listOf(NAME1))
+        assertThat(referee.decideWinner(cars)).isEqualTo(listOf(NAME1))
     }
 
     @Test
     fun `우승자가 여러 명일 경우`() {
         val cars = listOf(Car(NAME1, WINNER_POSITION), Car(NAME2, WINNER_POSITION))
-        assertThat(CarRacingGame().decideWinner(cars)).isEqualTo(listOf(NAME1, NAME2))
+        assertThat(referee.decideWinner(cars)).isEqualTo(listOf(NAME1, NAME2))
     }
 
     companion object {
